Move MacRadio and ProxyRadio to their own modules
