+++ openthread/src/radio.rs
@@ -4,31 +4,21 @@
 
 #![allow(clippy::unnecessary_cast)]
 
-use core::cell::Cell;
 use core::fmt::Debug;
-use core::future::Future;
-use core::mem::MaybeUninit;
-use core::pin::pin;
 
-use embassy_futures::select::{select, Either};
-
-use embassy_sync::blocking_mutex::raw::{CriticalSectionRawMutex, RawMutex};
-use embassy_sync::signal::Signal;
-use embassy_sync::zerocopy_channel::{Channel, Receiver, Sender};
-
-use embassy_time::Instant;
-
-use mac::MacHeader;
 use openthread_sys::{
     OT_RADIO_CAPS_ALT_SHORT_ADDR, OT_RADIO_CAPS_RX_ON_WHEN_IDLE, OT_RADIO_CAPS_TRANSMIT_FRAME_POWER,
 };
 
-use crate::fmt::{bitflags, Bytes};
+use crate::fmt::bitflags;
 use crate::sys::{
     OT_RADIO_CAPS_ACK_TIMEOUT, OT_RADIO_CAPS_CSMA_BACKOFF, OT_RADIO_CAPS_ENERGY_SCAN,
     OT_RADIO_CAPS_RECEIVE_TIMING, OT_RADIO_CAPS_SLEEP_TO_TX, OT_RADIO_CAPS_TRANSMIT_RETRIES,
-    OT_RADIO_CAPS_TRANSMIT_SEC, OT_RADIO_CAPS_TRANSMIT_TIMING, OT_RADIO_FRAME_MAX_SIZE,
+    OT_RADIO_CAPS_TRANSMIT_SEC, OT_RADIO_CAPS_TRANSMIT_TIMING,
 };
+
+pub use mac::*;
+pub use proxy::*;
 
 // Concrete [`Radio`] implementations for the supported radio hardware /
 // deployments. Each is gated on the feature that enables it.
@@ -40,8 +30,10 @@ use crate::sys::{
 //   wire protocol it speaks.)
 #[cfg(feature = "esp-radio")]
 pub mod esp;
+mod mac;
 #[cfg(feature = "embassy-nrf")]
 pub mod nrf;
+mod proxy;
 #[cfg(feature = "rcp")]
 pub mod spinel;
 
@@ -568,1362 +560,5 @@ where
 
     async fn receive(&mut self, psdu_buf: &mut [u8]) -> Result<PsduMeta, Self::Error> {
         T::receive(self, psdu_buf).await
-    }
-}
-
-/// An error type for the enhanced radio.
-#[derive(Debug, Copy, Clone, Eq, PartialEq, Hash)]
-pub enum MacRadioError<T> {
-    /// Invalid TX frame
-    TxInvalid,
-    /// Invalid RX frame
-    RxInvalid,
-    /// Receiving failed due to sending an ACK frame failed
-    TxAckFailed(T),
-    /// Transmitting failed due to receiving an ACK frame failed
-    RxAckFailed(T),
-    /// Receiving failed due to timeout when preparing an ACK frame
-    TxAckTimeout,
-    /// Transmitting failed due to no ACK received
-    RxAckTimeout,
-    /// Transmitting failed due to invalid ACK received
-    RxAckInvalid,
-    /// Error coming from the wrapped radio
-    Io(T),
-}
-
-impl<T> RadioError for MacRadioError<T>
-where
-    T: RadioError,
-{
-    fn kind(&self) -> RadioErrorKind {
-        match self {
-            Self::TxInvalid => RadioErrorKind::TxInvalid,
-            Self::RxInvalid => RadioErrorKind::RxInvalid,
-            Self::RxAckInvalid => RadioErrorKind::RxAckInvalid,
-            Self::TxAckFailed(_) => RadioErrorKind::TxAckFailed,
-            Self::RxAckFailed(_) => RadioErrorKind::RxAckFailed,
-            Self::TxAckTimeout => RadioErrorKind::TxAckTimeout,
-            Self::RxAckTimeout => RadioErrorKind::RxAckTimeout,
-            Self::Io(e) => e.kind(),
-        }
-    }
-}
-
-#[cfg(feature = "defmt")]
-impl<T> defmt::Format for MacRadioError<T>
-where
-    T: RadioError,
-{
-    fn format(&self, fmt: defmt::Formatter<'_>) {
-        defmt::write!(fmt, "{}", self.kind())
-    }
-}
-
-/// An enhanced (MAC) radio that can optionally send and receive ACKs for transmitted frames
-/// as well as optionally do address filtering.
-pub struct MacRadio<R, T> {
-    /// The wrapped radio.
-    radio: R,
-    /// The timer implementation to use.
-    /// Necessary to properly time:
-    /// - How long to wait for an ACK for a transmitted frame
-    /// - How long to wait before sending an ACK for a received frame
-    ///
-    /// The above is only relevant if the `MacRadio` is instructed to
-    /// receive TX ACKs in software and/or to send RX ACKs in software.
-    ///
-    /// Should be with a high precision of ideally < 10us.
-    timer: T,
-    /// The wrapped radio's MAC-offloading capabilities, discovered from its
-    /// `init`. Any capability the radio lacks, this wrapper emulates in software;
-    /// the `transmit`/`receive` paths branch on this at runtime (the wrapped
-    /// radio's MAC caps are only known after `init`, so they cannot be a const).
-    mac_caps: MacCapabilities,
-    /// Whether the radio is in promiscuous mode.
-    promiscuous: bool,
-    /// A buffer for the MAC header of the received or transmitted frame.
-    /// (For filtering and ACKs)
-    mac_header: MacHeader,
-    /// The buffer for the ACK PSDU, if the `MacRadio` is instructed
-    /// to send or receive ACKs in software.
-    // TODO: Inject from outside
-    ack_psdu_buf: [u8; OT_RADIO_FRAME_MAX_SIZE as _],
-    /// Frames accepted (and ACKed) while `transmit` was waiting for its own
-    /// ACK, parked here until subsequent `receive` calls deliver them. Sized
-    /// for a line-rate request burst: while this node serializes its replies
-    /// (each a full transmit sequence), further requests keep arriving and
-    /// park here - a depth of 12 absorbs the bursts the upstream CLI suites
-    /// fire (ten back-to-back pings) with headroom at real-time pacing,
-    /// where 4 measurably dropped the tail and 8 was marginal. Anything
-    /// beyond is dropped like on a saturated real radio.
-    // TODO: Inject from outside
-    pending_rx: heapless::Deque<(PsduMeta, [u8; OT_RADIO_FRAME_MAX_SIZE as _]), 12>,
-    /// The source-address-match table, consulted for the Frame Pending bit
-    /// of the software ACKs answering data polls (see [`SrcMatchEntries`]).
-    // TODO: Inject from outside
-    src_match: SrcMatchEntries,
-    /// The PAN ID to filter by, if the filter policy allows it.
-    pan_id: u16,
-    /// The short address to filter by, if the filter policy allows it.
-    short_addr: u16,
-    /// The alternate short address to *also* accept, or `BROADCAST_SHORT_ADDR`
-    /// (`0xffff`) when there is none. A real destination address never equals
-    /// the broadcast sentinel, so an unset alternate never matches — a no-op.
-    alt_short_addr: u16,
-    /// The extended address to filter by, if the filter policy allows it.
-    ext_addr: u64,
-}
-
-impl<R, T> MacRadio<R, T>
-where
-    R: Radio,
-    T: MacRadioTimer,
-{
-    /// The waiting timeout for a TX ACK to be received: OpenThread's own
-    /// SubMac ACK timeout (`kAckTimeout`, 16ms).
-    ///
-    /// Besides matching the stack's expectations, keeping this short bounds
-    /// a structural cost of this wrapper: a frame that crosses our
-    /// transmission is ACKed and parked during the wait (see `transmit`),
-    /// but reaches the stack only once the wait resolves - so on ACK loss,
-    /// this timeout is the worst-case added latency for the crossing
-    /// frame's processing (and thus for our reply to it). A real radio's
-    /// independent RX path has no such serialization.
-    const TX_ACK_WAIT_US: u64 = 16 * 1000;
-    /// The waiting timeout for an RX ACK to be sent.
-    // TODO: Should be 190us, but we need to be more precise
-    // and not use `embassy-time` with the NRF...
-    const RX_ACK_SEND_US: u64 = 10;
-
-    /// Create a new enhanced MAC radio.
-    ///
-    /// Arguments:
-    /// - `radio`: The radio to wrap.
-    /// - `delay`: The delay implementation to use. Should be with a high precision of ideally < 10us
-    /// - `ack_policy`: The ACK policy to use.
-    /// - `filter_policy`: The filter policy to use.
-    pub fn new(radio: R, timer: T) -> Self {
-        Self {
-            radio,
-            timer,
-            // Filled from the wrapped radio's `init`; until then assume no MAC
-            // offload (the wrapper emulates everything).
-            mac_caps: MacCapabilities::empty(),
-            mac_header: MacHeader::new(),
-            ack_psdu_buf: [0; OT_RADIO_FRAME_MAX_SIZE as _],
-            pending_rx: heapless::Deque::new(),
-            src_match: SrcMatchEntries::default(),
-            promiscuous: false,
-            pan_id: MacHeader::BROADCAST_PAN_ID,
-            short_addr: MacHeader::BROADCAST_SHORT_ADDR,
-            alt_short_addr: MacHeader::BROADCAST_SHORT_ADDR,
-            ext_addr: MacHeader::BROADCAST_EXT_ADDR,
-        }
-    }
-
-    /// Screen an incoming frame: apply the software address filters the
-    /// wrapped radio does not offload and - for an accepted frame that
-    /// requests one - send the ACK. Returns whether the frame is for us and
-    /// should be delivered to the stack.
-    ///
-    /// Shared by the `receive` path and by `transmit`'s ACK wait, so that a
-    /// frame crossing our transmission is served all the same (see there).
-    ///
-    /// `psdu` must not alias `self`'s buffers (callers pass caller-owned or
-    /// stack copies).
-    async fn screen_incoming(&mut self, psdu: &[u8]) -> Result<bool, MacRadioError<R::Error>> {
-        if self.mac_caps == MacCapabilities::all() {
-            return Ok(true);
-        }
-
-        if self.mac_header.load(psdu).is_none() {
-            trace!(
-                "MacRadio, received frame with invalid MAC header, dropping: {}",
-                Bytes(psdu)
-            );
-            return Ok(false);
-        }
-
-        if !self.mac_caps.contains(MacCapabilities::PROMISCUOUS) && !self.promiscuous {
-            if !self.mac_caps.contains(MacCapabilities::FILTER_PAN_ID)
-                && self.mac_header.pan_id != MacHeader::BROADCAST_PAN_ID
-                && self.mac_header.pan_id != self.pan_id
-            {
-                trace!(
-                    "MacRadio, filtering out frame: {}, PAN ID does not match",
-                    Bytes(psdu)
-                );
-                return Ok(false);
-            }
-
-            if !self.mac_caps.contains(MacCapabilities::FILTER_SHORT_ADDR)
-                && self.mac_header.dst_short_addr != MacHeader::BROADCAST_SHORT_ADDR
-                && self.mac_header.dst_short_addr != self.short_addr
-                && self.mac_header.dst_short_addr != self.alt_short_addr
-            {
-                trace!(
-                    "MacRadio, filtering out frame: {}, short address does not match",
-                    Bytes(psdu)
-                );
-                return Ok(false);
-            }
-
-            if !self.mac_caps.contains(MacCapabilities::FILTER_EXT_ADDR)
-                && self.mac_header.dst_ext_addr != MacHeader::BROADCAST_EXT_ADDR
-                && self.mac_header.dst_ext_addr != self.ext_addr
-            {
-                trace!(
-                    "MacRadio, filtering out frame: {}, extended address does not match",
-                    Bytes(psdu)
-                );
-                return Ok(false);
-            }
-
-            if !self.mac_caps.contains(MacCapabilities::RX_ACK) && self.mac_header.needs_ack() {
-                let ack_at = self.timer.now() + Self::RX_ACK_SEND_US;
-
-                // Ack MAC command frames with Frame Pending set: a
-                // sleepy child's data poll is a command frame, and the
-                // FP bit is what keeps the child awake for the queued
-                // indirect frame. Conservative - like the C simulation
-                // radio with source matching disabled: a spurious FP
-                // only keeps a child listening briefly when nothing
-                // is queued.
-                let frame_pending = self.mac_header.is_command()
-                    && self.src_match.ack_frame_pending(
-                        self.mac_header.src_short_addr,
-                        self.mac_header.src_ext_addr,
-                    );
-
-                let ack_len = self
-                    .mac_header
-                    .prep_ack(&mut self.ack_psdu_buf, frame_pending);
-
-                trace!(
-                    "MacRadio, about to transmit ACK: {}",
-                    Bytes(&self.ack_psdu_buf[..ack_len])
-                );
-
-                if self.timer.now() < ack_at {
-                    self.timer.wait(ack_at).await;
-                }
-
-                self.radio
-                    .transmit(&self.ack_psdu_buf[..ack_len], false, None)
-                    .await
-                    .map_err(MacRadioError::TxAckFailed)?;
-            }
-        }
-
-        Ok(true)
-    }
-}
-
-impl<R, T> Radio for MacRadio<R, T>
-where
-    R: Radio,
-    T: MacRadioTimer,
-{
-    type Error = MacRadioError<R::Error>;
-
-    async fn init(&mut self) -> Result<RadioCaps, Self::Error> {
-        let caps = self.radio.init().await.map_err(Self::Error::Io)?;
-
-        // Remember the wrapped radio's *actual* MAC offload so `transmit`/
-        // `receive` know what to emulate. To the outside, a `MacRadio` provides
-        // the full MAC-offload set — whatever the hardware doesn't do, this
-        // wrapper does in software — while the PHY caps pass through unchanged.
-        self.mac_caps = caps.mac;
-
-        // Outwards: the full MAC-offload set - whatever the hardware doesn't
-        // do, this wrapper does in software - with ONE exception: for an
-        // inner radio that sends its own RX ACKs but has no source-match
-        // table, `SRC_MATCH` cannot be claimed by anyone (the ACKs' Frame
-        // Pending bits are decided below, out of software's reach).
-        let mut mac = MacCapabilities::all();
-        if caps.mac.contains(MacCapabilities::RX_ACK)
-            && !caps.mac.contains(MacCapabilities::SRC_MATCH)
-        {
-            mac.remove(MacCapabilities::SRC_MATCH);
-        }
-
-        Ok(RadioCaps {
-            phy: caps.phy,
-            mac,
-            receive_sensitivity: caps.receive_sensitivity,
-        })
-    }
-
-    async fn set_config(&mut self, config: &Config) -> Result<(), Self::Error> {
-        self.radio
-            .set_config(config)
-            .await
-            .map_err(Self::Error::Io)?;
-
-        self.promiscuous = config.promiscuous;
-        self.pan_id = config.pan_id.unwrap_or(MacHeader::BROADCAST_PAN_ID);
-        self.short_addr = config.short_addr.unwrap_or(MacHeader::BROADCAST_SHORT_ADDR);
-        self.alt_short_addr = config
-            .alt_short_addr
-            .unwrap_or(MacHeader::BROADCAST_SHORT_ADDR);
-        self.ext_addr = config.ext_addr.unwrap_or(MacHeader::BROADCAST_EXT_ADDR);
-
-        Ok(())
-    }
-
-    async fn set_src_match(&mut self, entries: &SrcMatchEntries) -> Result<(), Self::Error> {
-        if self.mac_caps.contains(MacCapabilities::SRC_MATCH) {
-            // The inner radio's own acking honors the table - hand it down.
-            self.radio
-                .set_src_match(entries)
-                .await
-                .map_err(Self::Error::Io)
-        } else {
-            // This wrapper's software ACKs consult the copy.
-            self.src_match = entries.clone();
-
-            Ok(())
-        }
-    }
-
-    async fn energy_scan(&mut self, duration_millis: u16) -> Result<i8, Self::Error> {
-        // Energy scan involves no MAC-layer processing - pass through.
-        self.radio
-            .energy_scan(duration_millis)
-            .await
-            .map_err(Self::Error::Io)
-    }
-
-    async fn transmit(
-        &mut self,
-        psdu: &[u8],
-        cca: bool,
-        ack_psdu_buf: Option<&mut [u8]>,
-    ) -> Result<Option<PsduMeta>, Self::Error> {
-        trace!("MacRadio, about to transmit");
-
-        if self.mac_caps.contains(MacCapabilities::TX_ACK) {
-            let result = self
-                .radio
-                .transmit(psdu, cca, ack_psdu_buf)
-                .await
-                .map_err(Self::Error::Io);
-
-            trace!("MacRadio, transmitted");
-
-            result
-        } else {
-            self.radio
-                .transmit(psdu, cca, None)
-                .await
-                .map_err(Self::Error::Io)?;
-
-            let sent_at = self.timer.now();
-
-            self.mac_header.load(psdu).ok_or(MacRadioError::TxInvalid)?;
-
-            if self.mac_header.needs_ack() {
-                let psdu_seq = self.mac_header.seq;
-
-                trace!("MacRadio, about to receive transmit ACK");
-
-                // Wait for the matching ACK until the deadline, SERVING the
-                // medium meanwhile: on a receive-everything PHY, other frames
-                // routinely land in this window (a neighbor's broadcast, a
-                // crossing transmission), and they must be screened, ACKed
-                // and parked for `receive` just as if no wait were running -
-                // which is what a real radio's independent RX path (and the C
-                // simulation radio in TX wait) does. Merely dropping them
-                // makes two nodes retransmitting to each other mutually deaf
-                // - each sitting in its own ACK wait, ACKing nothing - until
-                // their MAC retry budgets run out.
-                let ack_meta = loop {
-                    let result = {
-                        let mut ack = pin!(self.radio.receive(&mut self.ack_psdu_buf));
-                        let mut timeout = pin!(self.timer.wait(sent_at + Self::TX_ACK_WAIT_US));
-
-                        select(&mut ack, &mut timeout).await
-                    };
-
-                    let meta = match result {
-                        Either::First(result) => result.map_err(Self::Error::RxAckFailed)?,
-                        Either::Second(_) => {
-                            trace!("MacRadio, transmit ACK timeout");
-
-                            Err(Self::Error::RxAckTimeout)?
-                        }
-                    };
-
-                    let psdu = &self.ack_psdu_buf[..meta.len];
-                    if self
-                        .mac_header
-                        .load(psdu)
-                        .is_some_and(|()| self.mac_header.ack_for(psdu_seq))
-                    {
-                        break meta;
-                    }
-
-                    // A crossing frame: screen it off a stack copy (the ACK
-                    // buffer is about to be reused for both the ACK we may
-                    // send and the wait's next read).
-                    let mut crossing = [0; OT_RADIO_FRAME_MAX_SIZE as _];
-                    crossing[..meta.len].copy_from_slice(&self.ack_psdu_buf[..meta.len]);
-
-                    if self.screen_incoming(&crossing[..meta.len]).await?
-                        && self.pending_rx.push_back((meta, crossing)).is_err()
-                    {
-                        trace!(
-                            "MacRadio, crossing-frame queue full, dropped: {}",
-                            Bytes(&crossing[..meta.len])
-                        );
-                    }
-                };
-
-                let ack_psdu = &self.ack_psdu_buf[..ack_meta.len];
-
-                if let Some(ack_psdu_buf) = ack_psdu_buf {
-                    ack_psdu_buf[..ack_psdu.len()].copy_from_slice(ack_psdu);
-                }
-
-                trace!("MacRadio, transmitted with ACK");
-
-                // Report the received ACK: the stack reads more than "it was
-                // acked" out of it - notably the Frame Pending bit, which
-                // tells a sleepy child whether to stay awake for a frame its
-                // parent has queued.
-                Ok(Some(ack_meta))
-            } else {
-                trace!("MacRadio, transmitted without ACK");
-
-                Ok(None)
-            }
-        }
-    }
-
-    async fn receive(&mut self, psdu_buf: &mut [u8]) -> Result<PsduMeta, Self::Error> {
-        // A frame screened - and already ACKed - during a transmission's ACK
-        // wait is delivered first.
-        if let Some((psdu_meta, psdu)) = self.pending_rx.pop_front() {
-            psdu_buf[..psdu_meta.len].copy_from_slice(&psdu[..psdu_meta.len]);
-
-            trace!(
-                "MacRadio, delivering frame parked during an ACK wait: {}",
-                Bytes(&psdu_buf[..psdu_meta.len])
-            );
-
-            return Ok(psdu_meta);
-        }
-
-        loop {
-            trace!("MacRadio, about to receive");
-
-            let psdu_meta = self
-                .radio
-                .receive(psdu_buf)
-                .await
-                .map_err(Self::Error::Io)?;
-
-            trace!(
-                "MacRadio, received: {}, meta: {:?}",
-                Bytes(&psdu_buf[..psdu_meta.len]),
-                psdu_meta
-            );
-
-            if self.screen_incoming(&psdu_buf[..psdu_meta.len]).await? {
-                trace!(
-                    "MacRadio, received frame: {}",
-                    Bytes(&psdu_buf[..psdu_meta.len])
-                );
-
-                break Ok(psdu_meta);
-            }
-        }
-    }
-}
-
-/// A high-res timer trait that is necessary for the `MacRadio` to send ACKs
-/// at the right time.
-///
-/// The timer should have a microsecond, or a few microseconds' resolution.
-pub trait MacRadioTimer {
-    /// Returns the current time - in microseconds - from some predefined period in time
-    /// The returned time should be monotonic.
-    fn now(&mut self) -> u64;
-
-    /// Waits until the current time becomes equal or greater than the given time.
-    ///
-    /// If the current time is already equal or greater than the given time,
-    /// the function should return immediately, without waiting.
-    ///
-    /// Arguments:
-    /// - `time`: The time - in microseconds - to wait until.
-    async fn wait(&mut self, at: u64);
-}
-
-impl<T> MacRadioTimer for &mut T
-where
-    T: MacRadioTimer,
-{
-    fn now(&mut self) -> u64 {
-        T::now(self)
-    }
-
-    async fn wait(&mut self, at: u64) {
-        T::wait(self, at).await
-    }
-}
-
-/// An implementation of `MacRadioTimer` that uses the `embassy_time` crate.
-///
-/// Note that this implementation might NOT be appropriate when the concrete
-/// `embassy-time` implementation is not having a high-enough resolution.
-pub struct EmbassyTimeTimer;
-
-impl MacRadioTimer for EmbassyTimeTimer {
-    fn now(&mut self) -> u64 {
-        Instant::now().as_micros()
-    }
-
-    async fn wait(&mut self, at: u64) {
-        embassy_time::Timer::at(Instant::from_micros(at)).await;
-    }
-}
-
-/// The resources for the radio proxy.
-pub struct ProxyRadioResources {
-    request_buf: MaybeUninit<[ProxyRadioRequest; 1]>,
-    response_buf: MaybeUninit<[ProxyRadioResponse; 1]>,
-    state: MaybeUninit<ProxyRadioState<'static>>,
-}
-
-impl ProxyRadioResources {
-    /// Create a new set of radio proxy resources.
-    pub const fn new() -> Self {
-        Self {
-            request_buf: MaybeUninit::uninit(),
-            response_buf: MaybeUninit::uninit(),
-            state: MaybeUninit::uninit(),
-        }
-    }
-}
-
-impl Default for ProxyRadioResources {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// A type that allows to offload the execution (TX/RX) of the actual PHY `Radio` impl
-/// to a separate - possibly higher-priority - executor.
-///
-/// Running the PHY radio in a separate higher priority executor is particularly desirable in the cases where it
-/// cannot do MAC-offloading (ACKs and filtering) in hardware, and hence the `MacRadio` wrapper is used to handle
-/// these tasks in software. Due to timing constraints with ACKs and filtering, this task should have a higher
-/// priority than all other `OpenThread`-related tasks.
-///
-/// This is achieved by splitting the radio into two types:
-/// - `ProxyRadio`, which is a radio proxy that implements the `Radio` trait and is to be used by the main execution
-///   by passing it to `OpenThread::run`
-/// - `PhyRadioRunner`, which is `Send` and therefore can be sent to a separate executor - to run the radio.
-///   Invoke `PhyRadioRunner::run(<the-phy-radio>, <delay-provider>).await` in that separate executor.
-pub struct ProxyRadio<'a> {
-    /// The request channel to the PHY radio
-    request: Sender<'a, CriticalSectionRawMutex, ProxyRadioRequest>,
-    /// The response channel from the PHY radio
-    response: Receiver<'a, CriticalSectionRawMutex, ProxyRadioResponse>,
-    /// Whether the current response was cancelled
-    /// If set to `true`, this means we have to drop the pending response from the driver
-    cancelled: Cell<bool>,
-    /// The signal to indicate a new request to the PHY radio, so that
-    /// the PHY radio can cancel the current request (if any) and start processing the new one
-    cancel: &'a Signal<CriticalSectionRawMutex, ()>,
-    /// The PHY radio's capabilities, published by the runner. `init` awaits it.
-    caps: &'a Signal<CriticalSectionRawMutex, RadioCaps>,
-    /// The current radio configuration
-    config: Config,
-}
-
-impl<'a> ProxyRadio<'a> {
-    const INIT_REQUEST: [ProxyRadioRequest; 1] = [ProxyRadioRequest::new()];
-    const INIT_RESPONSE: [ProxyRadioResponse; 1] = [ProxyRadioResponse::new()];
-
-    /// Create a new `ProxyRadio` and its `PhyRadioRunner` instances.
-    ///
-    /// Arguments:
-    /// - `resources`: The radio proxy resources
-    pub fn new(resources: &'a mut ProxyRadioResources) -> (Self, PhyRadioRunner<'a>) {
-        resources.request_buf.write(Self::INIT_REQUEST);
-        resources.response_buf.write(Self::INIT_RESPONSE);
-
-        let request_buf = unsafe { resources.request_buf.assume_init_mut() };
-        let response_buf = unsafe { resources.response_buf.assume_init_mut() };
-
-        resources.state.write(ProxyRadioState::new(
-            unsafe {
-                core::mem::transmute::<
-                    &mut [ProxyRadioRequest; 1],
-                    &'static mut [ProxyRadioRequest; 1],
-                >(request_buf)
-            },
-            unsafe {
-                core::mem::transmute::<
-                    &mut [ProxyRadioResponse; 1],
-                    &'static mut [ProxyRadioResponse; 1],
-                >(response_buf)
-            },
-        ));
-
-        let state = unsafe { resources.state.assume_init_mut() };
-
-        state.split()
-    }
-
-    /// Clear any cancelled response from the driver
-    async fn process_cancelled(&mut self) {
-        if self.cancelled.get() {
-            self.response.receive().await;
-
-            self.cancelled.set(false);
-            self.response.receive_done();
-
-            trace!("ProxyRadio, cancelled response cleared");
-        }
-    }
-}
-
-impl Radio for ProxyRadio<'_> {
-    type Error = RadioErrorKind;
-
-    async fn init(&mut self) -> Result<RadioCaps, Self::Error> {
-        // The actual radio (and its `init`) lives on the `PhyRadioRunner`, which
-        // publishes the discovered capabilities to the shared `caps` signal.
-        // Await them here so the proxy reports the real, runtime-discovered set —
-        // both PHY and MAC — just like every other radio. (The runner wraps the
-        // radio in a `MacRadio`, so the MAC set it publishes is the full offload
-        // set; the runner must be running for TX/RX to work at all, so this
-        // resolves once it has brought the radio up.)
-        //
-        // TODO: the proxy's request/response protocol carries only TX/RX, so
-        // `Radio::energy_scan` is *not* forwarded to the PHY radio: the proxy
-        // keeps the default implementation and reports "no measurement". This
-        // costs nothing today (no proxied radio can measure channel energy —
-        // see the `esp`/`nrf` notes); forward it through the protocol if one
-        // ever can.
-        Ok(self.caps.wait().await)
-    }
-
-    async fn set_config(&mut self, config: &Config) -> Result<(), Self::Error> {
-        // There is no separate command for updating the configuration
-        // The updated configuration is always valid for the next request
-        self.config = config.clone();
-        Ok(())
-    }
-
-    async fn transmit(
-        &mut self,
-        psdu: &[u8],
-        cca: bool,
-        ack_psdu_buf: Option<&mut [u8]>,
-    ) -> Result<Option<PsduMeta>, Self::Error> {
-        trace!("ProxyRadio, about to transmit: {}", Bytes(psdu));
-
-        self.process_cancelled().await;
-
-        {
-            let req = self.request.send().await;
-
-            req.tx = true;
-            req.config = self.config.clone();
-            req.cca = cca;
-            req.psdu.clear();
-            unwrap!(req.psdu.extend_from_slice(psdu));
-
-            trace!("ProxyRadio, transmit request sent: {:?}", req);
-
-            self.request.send_done();
-        }
-
-        self.cancelled.set(true);
-        let _guard = scopeguard::guard((), |_| {
-            if self.cancelled.get() {
-                trace!("ProxyRadio, transmit request cancelled");
-                self.cancel.signal(())
-            }
-        });
-
-        trace!("ProxyRadio, waiting for transmit response");
-
-        let resp = self.response.receive().await;
-
-        trace!("ProxyRadio, transmit response received: {:?}", resp);
-
-        let psdu_meta = (ack_psdu_buf.is_some() && !resp.psdu.is_empty()).then_some(PsduMeta {
-            len: resp.psdu.len(),
-            channel: resp.psdu_channel,
-            rssi: resp.psdu_rssi,
-        });
-
-        if let Some(ack_psdu_buf) = ack_psdu_buf {
-            if psdu_meta.is_some() {
-                ack_psdu_buf[..resp.psdu.len()].copy_from_slice(&resp.psdu);
-            } else {
-                ack_psdu_buf.fill(0);
-            }
-        }
-
-        let result = resp.result.map(|_| psdu_meta);
-
-        self.cancelled.set(false);
-        self.response.receive_done();
-
-        trace!("ProxyRadio, transmit response done");
-
-        result
-    }
-
-    async fn receive(&mut self, psdu_buf: &mut [u8]) -> Result<PsduMeta, Self::Error> {
-        trace!("ProxyRadio, about to receive");
-
-        self.process_cancelled().await;
-
-        {
-            let req = self.request.send().await;
-
-            req.tx = false;
-            req.config = self.config.clone();
-            req.psdu.clear();
-
-            trace!("ProxyRadio, receive request sent: {:?}", req);
-
-            self.request.send_done();
-        }
-
-        self.cancelled.set(true);
-        let _guard = scopeguard::guard((), |_| {
-            if self.cancelled.get() {
-                trace!("ProxyRadio, receive request cancelled");
-                self.cancel.signal(());
-            }
-        });
-
-        trace!("ProxyRadio, waiting for receive response");
-
-        let resp = self.response.receive().await;
-
-        trace!("ProxyRadio, receive response received: {:?}", resp);
-
-        match resp.result {
-            Ok(()) => {
-                let len = resp.psdu.len();
-                psdu_buf[..len].copy_from_slice(&resp.psdu);
-
-                let psdu_meta = PsduMeta {
-                    len,
-                    channel: resp.psdu_channel,
-                    rssi: resp.psdu_rssi,
-                };
-
-                self.cancelled.set(false);
-                self.response.receive_done();
-
-                Ok(psdu_meta)
-            }
-            Err(e) => {
-                self.cancelled.set(false);
-                self.response.receive_done();
-                Err(e)
-            }
-        }
-    }
-}
-
-/// A type modeling the running of the PHY radio - the other side of the `ProxyRadio` pipe.
-pub struct PhyRadioRunner<'a> {
-    /// The request channel from the proxy radio
-    request: Receiver<'a, CriticalSectionRawMutex, ProxyRadioRequest>,
-    /// The response channel to the proxy radio
-    response: Sender<'a, CriticalSectionRawMutex, ProxyRadioResponse>,
-    /// The signal to indicate a new request from the proxy radio
-    /// which means we have to cancel processing the current request (if any)
-    cancel: &'a Signal<CriticalSectionRawMutex, ()>,
-    /// The signal on which we publish the actual radio's capabilities (after
-    /// running its `init`) so the `ProxyRadio` half can report them.
-    caps: &'a Signal<CriticalSectionRawMutex, RadioCaps>,
-}
-
-impl PhyRadioRunner<'_> {
-    /// Run the PHY radio.
-    ///
-    /// Arguments:
-    /// - `radio`: The PHY radio to run.
-    /// - `delay`: The delay implementation to use.
-    pub async fn run<R, T>(&mut self, radio: R, delay: T) -> !
-    where
-        R: Radio,
-        T: MacRadioTimer,
-    {
-        let mut radio = MacRadio::new(radio, delay);
-
-        // Bring the radio up before serving requests and publish its capabilities
-        // to the `ProxyRadio` half (which is what the OpenThread stack queries).
-        // `radio` is a `MacRadio`, so the MAC set is the full offload set. This is
-        // the runtime caps handshake across the executor boundary: the actual
-        // radio lives here, so only here can its `init` run. On failure we publish
-        // a default (empty) set rather than leave the proxy's `init` waiting
-        // forever; the radio may still recover lazily on the first request.
-        let caps = match radio.init().await {
-            Ok(caps) => caps,
-            Err(e) => {
-                warn!("PhyRadioRunner, radio init failed: {:?}", e);
-                RadioCaps::default()
-            }
-        };
-        self.caps.signal(caps);
-
-        debug!("PhyRadioRunner, running");
-
-        loop {
-            {
-                let request = self.request.receive().await;
-
-                if Self::process(&mut radio, request, &mut self.response, self.cancel)
-                    .await
-                    .is_some()
-                {
-                    trace!("PhyRadioRunner, processing done");
-                } else {
-                    // Processing was cancelled by a new request, need to send an "interrupted" response
-
-                    let response = self.response.send().await;
-
-                    response.psdu.clear();
-                    response.result = Err(RadioErrorKind::Other);
-                    response.psdu_channel = 0;
-                    response.psdu_rssi = None;
-
-                    self.response.send_done();
-
-                    trace!("PhyRadioRunner, processing cancelled");
-                }
-
-                self.request.receive_done();
-            }
-        }
-    }
-
-    // Process a single request (TX or RX) by first updating the driver configuration
-    // (driver should skip that if the new configuration is the same as the current one),
-    // and then transmitting or receiving the frame.
-    //
-    // Updating the configuration, as well as the TX/RX operation might be cancelled at
-    // any moment, if a new request arrives.
-    async fn process<T>(
-        mut radio: T,
-        request: &mut ProxyRadioRequest,
-        response_sender: &mut Sender<'_, impl RawMutex, ProxyRadioResponse>,
-        cancel: &Signal<impl RawMutex, ()>,
-    ) -> Option<()>
-    where
-        T: Radio,
-    {
-        let response = Self::with_cancel(response_sender.send(), cancel).await?;
-
-        response.psdu.clear();
-        response.psdu_channel = 0;
-        response.psdu_rssi = None;
-
-        trace!("PhyRadioRunner, processing request: {:?}", request);
-
-        // Always first set the configuration relevant for the current TX/RX request
-        // The PHY driver should have intelligence to skip the configuration update if the new
-        // configuration is the same as the current one
-        let result = Self::with_cancel(radio.set_config(&request.config), cancel)
-            .await?
-            .map_err(|e| e.kind());
-
-        trace!("PhyRadioRunner, configuration set: {:?}", result);
-
-        let result = if result.is_err() {
-            // Setting driver configuration resulted in an error, so skip the rest of the processing
-            result
-        } else {
-            unwrap!(response.psdu.resize_default(response.psdu.capacity()));
-
-            let result = if request.tx {
-                Self::with_cancel(
-                    radio.transmit(&request.psdu, request.cca, Some(&mut response.psdu)),
-                    cancel,
-                )
-                .await?
-                .map_err(|e| e.kind())
-            } else {
-                Self::with_cancel(radio.receive(&mut response.psdu), cancel)
-                    .await?
-                    .map_err(|e| e.kind())
-                    .map(Some)
-            };
-
-            if let Ok(Some(psdu_meta)) = &result {
-                response.psdu.truncate(psdu_meta.len);
-                response.psdu_channel = psdu_meta.channel;
-                response.psdu_rssi = psdu_meta.rssi;
-            } else {
-                // No frame returned, so clear the response fields
-                response.psdu.clear();
-            }
-
-            result.map(|_| ())
-        };
-
-        response.result = result;
-
-        trace!("PhyRadioRunner, processed response: {:?}", response);
-
-        response_sender.send_done();
-
-        Some(())
-    }
-
-    async fn with_cancel<F>(fut: F, cancel: &Signal<impl RawMutex, ()>) -> Option<F::Output>
-    where
-        F: Future,
-    {
-        match select(fut, cancel.wait()).await {
-            Either::First(result) => Some(result),
-            Either::Second(_) => None,
-        }
-    }
-}
-
-// Should be safe because while not (yet) marked formally as such, zerocopy-channel's
-// `Receiver` and `Sender` are `Send`, as long as the critical section is `Send` + `Sync`
-// (which is the case as we use `CriticalSectionRawMutex`), and the `ProxyRadioRequest` and
-// `ProxyRadioResponse` are `Send` (which is the case).
-//
-// The signals are obviously `Send` + `Sync`.
-unsafe impl Send for PhyRadioRunner<'_> {}
-
-const PSDU_LEN: usize = OT_RADIO_FRAME_MAX_SIZE as _;
-
-/// The state of the proxy radio
-///
-/// This state is borrowed and shared between
-/// the two ends of the pipe: the proxy radio, and the PHY radio runner.
-struct ProxyRadioState<'a> {
-    /// The request channel to the PHY radio
-    request: Channel<'a, CriticalSectionRawMutex, ProxyRadioRequest>,
-    /// The response channel from the PHY radio
-    response: Channel<'a, CriticalSectionRawMutex, ProxyRadioResponse>,
-    /// The signal to indicate a new request to the PHY radio
-    cancel: Signal<CriticalSectionRawMutex, ()>,
-    /// The PHY radio's capabilities, published by the runner (which owns the
-    /// actual radio and runs its `init`) once, and awaited by `ProxyRadio::init`.
-    /// This is how the proxy learns the caps at runtime instead of baking them
-    /// in — the actual radio lives on the runner's (possibly separate) executor.
-    caps: Signal<CriticalSectionRawMutex, RadioCaps>,
-}
-
-impl<'a> ProxyRadioState<'a> {
-    /// Create a new proxy radio state.
-    ///
-    /// Arguments:
-    /// - `request_buf`: The request buffer
-    /// - `response_buf`: The response buffer
-    fn new(
-        request_buf: &'a mut [ProxyRadioRequest; 1],
-        response_buf: &'a mut [ProxyRadioResponse; 1],
-    ) -> Self {
-        Self {
-            request: Channel::new(request_buf),
-            response: Channel::new(response_buf),
-            cancel: Signal::new(),
-            caps: Signal::new(),
-        }
-    }
-
-    /// Split the state into the proxy radio and the PHY radio runner.
-    fn split(&mut self) -> (ProxyRadio<'_>, PhyRadioRunner<'_>) {
-        let (request_sender, request_receiver) = self.request.split();
-        let (response_sender, response_receiver) = self.response.split();
-
-        (
-            ProxyRadio {
-                request: request_sender,
-                response: response_receiver,
-                cancelled: Cell::new(false),
-                cancel: &self.cancel,
-                caps: &self.caps,
-                config: Config::new(),
-            },
-            PhyRadioRunner {
-                request: request_receiver,
-                response: response_sender,
-                cancel: &self.cancel,
-                caps: &self.caps,
-            },
-        )
-    }
-}
-
-/// A proxy radio request.
-#[derive(Debug)]
-#[cfg_attr(feature = "defmt", derive(defmt::Format))]
-struct ProxyRadioRequest {
-    /// Transmit or receive
-    tx: bool,
-    /// Whether to perform clear channel assessment (CCA) before transmitting the frame
-    /// (only relevant for TX requests, should be ignored for RX requests)
-    cca: bool,
-    /// The radio configuration for the TX/RX operation
-    config: Config,
-    /// The PSDU to transmit for the TX operation
-    psdu: heapless::Vec<u8, PSDU_LEN>,
-}
-
-impl ProxyRadioRequest {
-    /// Create a new empty proxy radio request.
-    const fn new() -> Self {
-        Self {
-            tx: false,
-            cca: false,
-            config: Config::new(),
-            psdu: heapless::Vec::new(),
-        }
-    }
-}
-
-/// A proxy radio response.
-#[derive(Debug)]
-#[cfg_attr(feature = "defmt", derive(defmt::Format))]
-struct ProxyRadioResponse {
-    /// The result of the TX/RX operation
-    result: Result<(), RadioErrorKind>,
-    /// The received PSDU, if the operation was successful:
-    /// - For TX: the received ACK PSDU (might be empty)
-    /// - For RX: the received frame PSDU
-    psdu: heapless::Vec<u8, PSDU_LEN>,
-    /// The channel on which the frame was received:
-    /// - For TX: the channel on which the ACK frame was received
-    /// - For RX: the channel on which the regular frame was received
-    psdu_channel: u8,
-    /// The RSSI of the received frame, if the radio supports appending it at the end of the frame:
-    /// - For TX: the RSSI of the received ACK frame
-    /// - For RX: the RSSI of the received frame
-    psdu_rssi: Option<i8>,
-}
-
-impl ProxyRadioResponse {
-    /// Create a new empty proxy radio response.
-    const fn new() -> Self {
-        Self {
-            result: Ok(()),
-            psdu: heapless::Vec::new(),
-            psdu_channel: 0,
-            psdu_rssi: None,
-        }
-    }
-}
-
-/// A minimal set of utilities for parsing the IEEE 802.15.4 MAC header
-/// for the purposes of MAC filtering and RX/TX ACK processing.
-pub(crate) mod mac {
-    /// A parsed IEEE 802.15.4 MAC header.
-    pub struct MacHeader {
-        /// Frame Control Field (FCF)
-        pub fcf: u16,
-        /// Sequence number
-        pub seq: u8,
-        /// PAN ID. 0xffff if the Frame does not contain a PAN ID
-        /// or if the PAN ID is the broadcast PAN ID
-        pub pan_id: u16,
-        /// Destination short address
-        /// 0xffff if the Frame does not contain a short address
-        /// or if the short address is the broadcast short address
-        pub dst_short_addr: u16,
-        /// Destination extended address
-        /// 0xffffffffffffffff if the Frame does not contain an extended address
-        /// or if the extended address is the broadcast extended address
-        pub dst_ext_addr: u64,
-        /// Source short address
-        /// 0xffff if the Frame does not carry a short source address
-        pub src_short_addr: u16,
-        /// Source extended address
-        /// 0xffffffffffffffff if the Frame does not carry an extended source
-        /// address
-        pub src_ext_addr: u64,
-    }
-
-    impl MacHeader {
-        /// The length of an Imm-ACK PSDU.
-        pub const ACK_PSDU_LEN: usize = Self::FCF_LEN + Self::SEQ_LEN + Self::CRC_LEN;
-
-        /// The broadcast PAN ID.
-        pub const BROADCAST_PAN_ID: u16 = u16::MAX;
-        /// The broadcast short address.
-        pub const BROADCAST_SHORT_ADDR: u16 = u16::MAX;
-        /// The broadcast extended address.
-        pub const BROADCAST_EXT_ADDR: u64 = u64::MAX;
-
-        const FCF_LEN: usize = 2;
-        const SEQ_LEN: usize = 1;
-        const CRC_LEN: usize = 2;
-
-        const FCF_OFFSET: usize = 0;
-        const SEQ_OFFSET: usize = Self::FCF_LEN;
-        const ADDRS_OFFSET: usize = Self::SEQ_OFFSET + Self::SEQ_LEN;
-
-        const FCF_FRAME_TYPE_MASK: u16 = 0x07;
-        const FCF_FRAME_TYPE_ACK: u16 = 0x02;
-        #[allow(unused)]
-        const FCF_SECURITY_BIT: u16 = 1 << 3;
-        #[allow(unused)]
-        const FCF_PENDING_BIT: u16 = 1 << 4;
-        const FCF_ACK_REQ_BIT: u16 = 1 << 5;
-        #[allow(unused)]
-        const FCF_PAN_ID_COMPRESSION_MASK: u16 = 1 << 6;
-        const FCF_FRAME_DST_ADDR_MODE_SHIFT: u16 = 10;
-        const FCF_FRAME_DST_ADDR_MODE_MASK: u16 = 0x03 << Self::FCF_FRAME_DST_ADDR_MODE_SHIFT;
-        const FCF_FRAME_VERSION_SHIFT: u16 = 12;
-        const FCF_FRAME_VERSION_MASK: u16 = 0x03 << Self::FCF_FRAME_VERSION_SHIFT;
-        const FCF_FRAME_SRC_ADDR_MODE_SHIFT: u16 = 14;
-        const FCF_FRAME_SRC_ADDR_MODE_MASK: u16 = 0x03 << Self::FCF_FRAME_SRC_ADDR_MODE_SHIFT;
-
-        /// Create a new empty MAC header.
-        pub const fn new() -> Self {
-            Self {
-                fcf: 0,
-                seq: 0,
-                pan_id: 0,
-                dst_short_addr: 0,
-                dst_ext_addr: 0,
-                src_short_addr: 0,
-                src_ext_addr: 0,
-            }
-        }
-
-        /// Load the MAC header from a PSDU.
-        /// Returns `Some(())` if the MAC header was successfully loaded.
-        ///
-        /// This method will fail if the frame version or type is unknown (reserved)
-        /// or if the PSDU is too short.
-        #[inline(always)]
-        pub fn load(&mut self, psdu: &[u8]) -> Option<()> {
-            Self::ensure_len(psdu, Self::ADDRS_OFFSET + Self::CRC_LEN)?;
-
-            self.fcf =
-                u16::from_le_bytes(unwrap!(psdu[Self::FCF_OFFSET..Self::SEQ_OFFSET].try_into()));
-            self.seq = psdu[Self::SEQ_OFFSET];
-
-            let _frame_type = FrameType::get(self.fcf)?;
-            let _frame_version = FrameVersion::get(self.fcf)?;
-            let dst_addr_mode = FrameAddrMode::get_dst(self.fcf)?;
-
-            match dst_addr_mode {
-                FrameAddrMode::NotPresent => {
-                    self.pan_id = Self::BROADCAST_PAN_ID;
-                    self.dst_short_addr = Self::BROADCAST_SHORT_ADDR;
-                    self.dst_ext_addr = Self::BROADCAST_EXT_ADDR;
-                }
-                FrameAddrMode::Short => {
-                    Self::ensure_len(psdu, Self::ADDRS_OFFSET + 2 + 2 + Self::CRC_LEN)?;
-
-                    self.pan_id = u16::from_le_bytes(unwrap!(psdu[3..5].try_into()));
-                    self.dst_short_addr = u16::from_le_bytes(unwrap!(psdu[5..7].try_into()));
-                    self.dst_ext_addr = Self::BROADCAST_EXT_ADDR;
-                }
-                FrameAddrMode::Extended => {
-                    Self::ensure_len(psdu, Self::ADDRS_OFFSET + 2 + 8 + Self::CRC_LEN)?;
-
-                    self.pan_id = u16::from_le_bytes(unwrap!(psdu[3..5].try_into()));
-                    // See platform.rs, `otPlatRadioSetExtendedAddress` impl
-                    self.dst_ext_addr = u64::from_le_bytes(unwrap!(psdu[5..13].try_into()));
-                    self.dst_short_addr = Self::BROADCAST_SHORT_ADDR;
-                }
-            }
-
-            let src_addr_mode = FrameAddrMode::get_src(self.fcf)?;
-
-            // Offset just past the destination addressing fields.
-            let mut offs = Self::ADDRS_OFFSET
-                + match dst_addr_mode {
-                    FrameAddrMode::NotPresent => 0,
-                    FrameAddrMode::Short => 2 + 2,
-                    FrameAddrMode::Extended => 2 + 8,
-                };
-
-            // The source PAN ID is elided when the PAN ID Compression bit is
-            // set - the common case for intra-PAN traffic, data polls
-            // included. (The elision rule modeled here is the 2003/2006 one;
-            // the 2015 frame version's table-based rules are not - the
-            // consumers of the source fields only look at 2006-era command
-            // frames.)
-            if !matches!(src_addr_mode, FrameAddrMode::NotPresent)
-                && (self.fcf & Self::FCF_PAN_ID_COMPRESSION_MASK) == 0
-            {
-                offs += 2;
-            }
-
-            match src_addr_mode {
-                FrameAddrMode::NotPresent => {
-                    self.src_short_addr = Self::BROADCAST_SHORT_ADDR;
-                    self.src_ext_addr = Self::BROADCAST_EXT_ADDR;
-                }
-                FrameAddrMode::Short => {
-                    Self::ensure_len(psdu, offs + 2 + Self::CRC_LEN)?;
-
-                    self.src_short_addr =
-                        u16::from_le_bytes(unwrap!(psdu[offs..offs + 2].try_into()));
-                    self.src_ext_addr = Self::BROADCAST_EXT_ADDR;
-                }
-                FrameAddrMode::Extended => {
-                    Self::ensure_len(psdu, offs + 8 + Self::CRC_LEN)?;
-
-                    self.src_ext_addr =
-                        u64::from_le_bytes(unwrap!(psdu[offs..offs + 8].try_into()));
-                    self.src_short_addr = Self::BROADCAST_SHORT_ADDR;
-                }
-            }
-
-            Some(())
-        }
-
-        /// Return `true` if the frame needs an ACK.
-        #[inline(always)]
-        pub fn needs_ack(&self) -> bool {
-            (self.fcf & Self::FCF_ACK_REQ_BIT) != 0
-        }
-
-        /// Return `true` if the frame is a MAC command frame.
-        ///
-        /// A sleepy child's Data Request (data poll) is one; its command ID
-        /// sits in the (potentially encrypted) MAC payload, so the frame
-        /// *type* is as precise as header-level parsing can get.
-        #[inline(always)]
-        pub fn is_command(&self) -> bool {
-            matches!(FrameType::get(self.fcf), Some(FrameType::Command))
-        }
-
-        /// Prepare an ACK PSDU.
-        /// Assumes that the parsed frame header indicates that ACK is necessary (`self.needs_ack` returns `true`)
-        ///
-        /// `frame_pending` sets the ACK's Frame Pending bit - the "stay awake,
-        /// data follows" hint a parent gives a polling sleepy child.
-        #[inline(always)]
-        pub fn prep_ack(&self, ack_buf: &mut [u8], frame_pending: bool) -> usize {
-            assert!(ack_buf.len() >= Self::ACK_PSDU_LEN);
-
-            let ack_fcf = Self::FCF_FRAME_TYPE_ACK
-                | (self.fcf & Self::FCF_FRAME_VERSION_MASK)
-                | if frame_pending {
-                    Self::FCF_PENDING_BIT
-                } else {
-                    0
-                };
-
-            ack_buf[0] = ack_fcf.to_le_bytes()[0];
-            ack_buf[1] = ack_fcf.to_le_bytes()[1];
-            ack_buf[2] = self.seq;
-            ack_buf[3] = 0; // CRC, will be filled-in by the PHY driver
-            ack_buf[4] = 0; // CRC, will be filled-in by the PHY driver
-
-            Self::ACK_PSDU_LEN
-        }
-
-        /// Return `true` if the frame is an ACK frame and is an ACK for the given source sequence number.
-        #[inline(always)]
-        pub fn ack_for(&self, src_seq: u8) -> bool {
-            matches!(unwrap!(FrameType::get(self.fcf)), FrameType::Ack) && src_seq == self.seq
-        }
-
-        #[inline(always)]
-        fn ensure_len(psdu: &[u8], len: usize) -> Option<()> {
-            (psdu.len() >= len).then_some(())
-        }
-    }
-
-    /// The supported IEEE 802.15.4 frame versions
-    #[derive(Debug)]
-    #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-    enum FrameVersion {
-        IEEE802154_2003,
-        IEEE802154_2006,
-    }
-
-    impl FrameVersion {
-        /// Get the frame version from the FCF.
-        ///
-        /// If the version is not supported, returns `None`.
-        #[inline(always)]
-        fn get(fcf: u16) -> Option<Self> {
-            match (fcf & MacHeader::FCF_FRAME_VERSION_MASK) >> MacHeader::FCF_FRAME_VERSION_SHIFT {
-                0 => Some(Self::IEEE802154_2003),
-                1 => Some(Self::IEEE802154_2006),
-                _ => None,
-            }
-        }
-    }
-
-    /// The supported IEEE 802.15.4 frame types
-    #[derive(Debug)]
-    #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-    enum FrameType {
-        Beacon,
-        Data,
-        Ack,
-        Command,
-    }
-
-    impl FrameType {
-        /// Get the frame type from the FCF.
-        ///
-        /// If the type is not supported, returns `None`.
-        #[inline(always)]
-        fn get(fcf: u16) -> Option<Self> {
-            match fcf & MacHeader::FCF_FRAME_TYPE_MASK {
-                0 => Some(Self::Beacon),
-                1 => Some(Self::Data),
-                2 => Some(Self::Ack),
-                3 => Some(Self::Command),
-                _ => None,
-            }
-        }
-    }
-
-    /// The supported IEEE 802.15.4 frame address modes
-    #[derive(Debug)]
-    #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-    enum FrameAddrMode {
-        NotPresent,
-        Short,
-        Extended,
-    }
-
-    impl FrameAddrMode {
-        /// Get the destination address mode from the FCF.
-        ///
-        /// If the mode is not supported, returns `None`.
-        #[inline(always)]
-        fn get_dst(fcf: u16) -> Option<Self> {
-            match (fcf & MacHeader::FCF_FRAME_DST_ADDR_MODE_MASK)
-                >> MacHeader::FCF_FRAME_DST_ADDR_MODE_SHIFT
-            {
-                0 => Some(Self::NotPresent),
-                2 => Some(Self::Short),
-                3 => Some(Self::Extended),
-                _ => None,
-            }
-        }
-
-        fn get_src(fcf: u16) -> Option<Self> {
-            match (fcf & MacHeader::FCF_FRAME_SRC_ADDR_MODE_MASK)
-                >> MacHeader::FCF_FRAME_SRC_ADDR_MODE_SHIFT
-            {
-                0 => Some(Self::NotPresent),
-                2 => Some(Self::Short),
-                3 => Some(Self::Extended),
-                _ => None,
-            }
-        }
     }
 }
